requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.2.0
class-registry==2.1.2
filters==1.3.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.10
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
six==1.16.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
yarl==1.9.2
